+++ index.js
@@ -1,9 +1,7 @@
 /*
- * Copyright (c) 2012-2014 Digital Bazaar, Inc. All rights reserved.
+ * Copyright (c) 2012-2015 Digital Bazaar, Inc. All rights reserved.
  */
 var bedrock = require('./lib/bedrock');
 
-// running in development mode
-// load dev config and start
-require('./configs/dev');
+// run with development defaults
 bedrock.start();